refact: 서비스 계층 구조 통일
